requirements.txt
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.20
tavily-python>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
chromadb>=0.4.0
pytest>=7.4.0
fastapi>=0.104.0
uvicorn>=0.24.0